Move stats collection to PrestoSparkTaskExecutorFactory
