+++ presto-spark-classloader-interface/src/main/java/com/facebook/presto/spark/classloader_interface/IPrestoSparkTaskExecutorFactory.java
@@ -16,6 +16,9 @@ package com.facebook.presto.spark.classloader_interface;
 import org.apache.spark.util.CollectionAccumulator;
 import scala.collection.Iterator;
 
+import java.util.List;
+import java.util.Map;
+
 public interface IPrestoSparkTaskExecutorFactory
 {
     <T extends PrestoSparkTaskOutput> IPrestoSparkTaskExecutor<T> create(
@@ -26,5 +29,6 @@ public interface IPrestoSparkTaskExecutorFactory
             PrestoSparkTaskInputs inputs,
             CollectionAccumulator<SerializedTaskInfo> taskInfoCollector,
             CollectionAccumulator<PrestoSparkShuffleStats> shuffleStatsCollector,
+            CollectionAccumulator<List<Map<String, String>>> genericShuffleStatsCollector,
             Class<T> outputType);
 }
